Decouple DefaultMessageProcessingBehaviour from InvokeHandlerBehaviour
The responsibility of handling the `SerializationException` in the `DefaultMessageProcessingBehaviour`
is only there because the `InvokeHandlerBehaviour` might throw the exception as part of its
activities.

On that bassis, the InvokeHandlerBehaviour should own that exception handling responsibility, rather
than something higher up the pipeline
