+++ Pat.Subscriber.UnitTests/Behaviours/InvokeHandlerBehaviour.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Text;
 using System.Threading.Tasks;
+using log4net;
 using Microsoft.Azure.ServiceBus;
 using Newtonsoft.Json;
 using NSubstitute;
@@ -26,18 +27,20 @@ namespace Pat.Subscriber.UnitTests.Behaviours
         {
         }
 
+        private readonly ILog _log = Substitute.For<ILog>();
+        private readonly SubscriberConfiguration _config = Substitute.For<SubscriberConfiguration>();
 
         [Fact]
-        public async Task  When_InvokedHandlerThrowsException_ThenBehaviourThrowsUnwrappedException()
+        public async Task When_InvokedHandlerThrowsException_ThenBehaviourThrowsUnwrappedException()
         {
-            var invokeHandlerBehaviour = Substitute.ForPartsOf<InvokeHandlerBehaviour>();
+            var invokeHandlerBehaviour = Substitute.ForPartsOf<InvokeHandlerBehaviour>(_log, _config);
             invokeHandlerBehaviour.GetHandlerForMessageType(Arg.Any<Message>())
                 .ReturnsForAnyArgs(new MessageTypeMapping(typeof(TestEvent), typeof(FakeHandler)));
 
             var messageDependencyScope = Substitute.For<IMessageDependencyScope>();
             messageDependencyScope.GetService(null).ReturnsForAnyArgs(new FakeHandler());
             messageDependencyScope.GetService<IMessageDeserialiser>().Returns(new NewtonsoftMessageDeserialiser());
-           var messageContext = new MessageContext
+            var messageContext = new MessageContext
             {
                 Message = new Message(Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(new TestEvent()))),
                 DependencyScope = messageDependencyScope
