+++ Pat.Subscriber/MessageProcessing/InvokeHandlerBehaviour.cs
@@ -5,6 +5,7 @@ using System.Runtime.ExceptionServices;
 using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
+using log4net;
 using Microsoft.Azure.ServiceBus;
 using Microsoft.Azure.ServiceBus.InteropExtensions;
 using Pat.Subscriber.Deserialiser;
@@ -18,25 +19,40 @@ namespace Pat.Subscriber.MessageProcessing
     /// </summary>
     public class InvokeHandlerBehaviour : IMessageProcessingBehaviour
     {
+        private readonly ILog _log;
+        private readonly SubscriberConfiguration _config;
+
+        public InvokeHandlerBehaviour(ILog log, SubscriberConfiguration config)
+        {
+            _log = log;
+            _config = config;
+        }
 
         public async Task Invoke(Func<MessageContext, Task> next, MessageContext messageContext)
         {
             var message = messageContext.Message;
-            var messageBody = await GetMessageBody(message).ConfigureAwait(false);
-
-            var handlerForMessageType = GetHandlerForMessageType(message);
-            var messageDeserialiser = messageContext.DependencyScope.GetService<IMessageDeserialiser>();
-            var messageHandler = messageContext.DependencyScope.GetService(handlerForMessageType.HandlerType);
-            var typedMessage = messageDeserialiser.DeserialiseObject(messageBody, handlerForMessageType.MessageType);
-
             try
             {
+                var messageBody = await GetMessageBody(message).ConfigureAwait(false);
+                var handlerForMessageType = GetHandlerForMessageType(message);
+                var messageDeserialiser = messageContext.DependencyScope.GetService<IMessageDeserialiser>();
+                var messageHandler = messageContext.DependencyScope.GetService(handlerForMessageType.HandlerType);
+                var typedMessage = messageDeserialiser.DeserialiseObject(messageBody, handlerForMessageType.MessageType);
+            
                 var handlerTask = (Task)handlerForMessageType.HandlerMethod.Invoke(messageHandler, new[] { typedMessage });
                 await handlerTask.ConfigureAwait(false);
+                await next(messageContext);
             }
             catch (TargetInvocationException exception)
             {
                 ExceptionDispatchInfo.Capture(exception.InnerException).Throw();
+            }
+            catch (SerializationException ex)
+            {
+                var messageType = GetMessageType(message);
+                var correlationId = GetCollelationId(message);
+                await messageContext.MessageReceiver.DeadLetterAsync(message.SystemProperties.LockToken).ConfigureAwait(false);
+                _log.Warn($"Unable to deserialise message body, message deadlettered. `{messageType}` correlation id `{correlationId}` on subscriber `{_config.SubscriberName}`.", ex);
             }
         }
 
@@ -48,16 +64,16 @@ namespace Pat.Subscriber.MessageProcessing
         }
 
 
-        private static async Task<string> GetMessageBody(Message message)
+        private async Task<string> GetMessageBody(Message message)
         {
             /*
-             * This is a work around to cope with two clients, one client 
-             * is publishing with a stream payload, the other is with a 
-             * string payload. Once all versions of the Pat.Subscriber.PatSender 
-             * are on 2.0.33 or above this can be removed.
-             * 
-             * Yes, this is awful.
-             */
+                * This is a work around to cope with two clients, one client 
+                * is publishing with a stream payload, the other is with a 
+                * string payload. Once all versions of the Pat.Subscriber.PatSender 
+                * are on 2.0.33 or above this can be removed.
+                * 
+                * Yes, this is awful.
+                */
             try
             {
                 return MessageAsStringReaderStrategy(message.Clone());
@@ -73,6 +89,20 @@ namespace Pat.Subscriber.MessageProcessing
                     return MessageAsUtf8ByteArrayReaderStrategy(message.Clone());
                 }
             }
+        }
+
+        private static string GetMessageType(Message message)
+        {
+            return message.UserProperties.ContainsKey("MessageType")
+                ? message.UserProperties["MessageType"].ToString()
+                : "Unknown Message Type";
+        }
+
+        private static string GetCollelationId(Message message)
+        {
+            return message.UserProperties.ContainsKey("PBCorrelationId")
+                ? message.UserProperties["PBCorrelationId"].ToString()
+                : "null";
         }
 
         private static string MessageAsStringReaderStrategy(Message message)
